added new functionalities to the event processor
This is a change done by Milton and me 2 years ago in a remote pair-programming session, to improve the event processor of Roassal.
I don't remember why we didn't merge it, nor create a PR... but I am recovering the change via cherrypicking as it is not possible to create a github PR since migration from object-profile/Roassal3 repo to this pharo-graphics/Roassal resulted in disconnected histories.