Simplify roassal baseline and fix problem in other version than P11
The baseline of roassal was redefining everything for every version of Pharo. I propose to have a common one and then to only add the specificities for each version of Pharo. 
Also this fixes a bug in Pharo 8, 9, 10, 12 that was loading some packages that were supposed to be only for P11.